examples/nova_agents/requirements.txt
nova-act
strands-agents>=1.2.0
strands-agents-tools
strands-nova
